requirements.txt
spacy>=3.7.0
requests>=2.31.0
scikit-learn>=1.5.0
numpy>=1.26.0
python-telegram-bot>=21.0
pyaudio>=0.2.14
openai-whisper>=20230124